arbitration-art-django/requirements/base.txt
django>=6.0,<6.1
djangorestframework>=3.17,<4.0
djangorestframework-simplejwt>=5.4,<6.0
django-cors-headers>=4.9,<5.0
django-environ>=0.13,<1.0

# PostgreSQL
psycopg[binary]>=3.2,<4.0
